+++ FizzBuzz2/Fizzbuzz.Domain/FizzbuzzResult.cs
@@ -3,8 +3,8 @@ namespace FizzBuzz.Domain
     public class FizzbuzzResult
     {
         public List<string> Results { get; set; } = new List<string>();
-        public int Start { get; set; }
-        public int End { get; set; }
+        public long Start { get; set; }
+        public long End { get; set; }
 
         public int FizzCount { get; set; }
         public int BuzzCount { get; set; }
